Abstract the bytes impl of Instruction
